pyproject.toml
[project]
name = "skill-learn-deep-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
    "tavily-python>=0.7.10",
]
